resolve rebase, use edit team component
